% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{.init}
\alias{.init}
\title{Initialise}
\usage{
.init(self, dom_id, offset = NULL, horizontal = FALSE, id = NULL)
}
\arguments{
\item{self}{A \code{waypoint} object.}

\item{dom_id}{Id of element to attach waypoint to.}

\item{offset}{Vertical offset, an integer (pixels) or a percentage passed as a string e.g. \code{"10"}.}

\item{horizontal}{Whether to apply waypoints of horizontal offset.}

\item{id}{Id of waypoint, see details.}
}
\description{
Initialise a waypoint
}
\details{
If you do not provide an \code{id} then you will have to rely on the \code{get_*} family of
of methods to get the callbacks. If you do, you can still rely on the latter but also may access 
the callbacks the traditional way with `input$idName_direction`, `input$idName_next`, and 
`input$idName_previous`.
}
