% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waypoints.R
\name{use_waypointer}
\alias{use_waypointer}
\title{Waypoints dependencies}
\usage{
use_waypointer()
}
\description{
Sources waypointer dependencies.
}
\details{
Place anywhere in your shiny UI.
}
\seealso{
\code{\link{waypoints}}
}
