% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrolly.R
\docType{data}
\name{Waypoint}
\alias{Waypoint}
\title{Waypoint}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Waypoint
}
\description{
A waypoint object to track.
}
\keyword{datasets}
