% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waypoints.R
\docType{data}
\name{waypoints}
\alias{waypoints}
\alias{Waypoint}
\title{Waypoint}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Waypoint
}
\description{
A waypoint object to track.
}
\section{Methods}{

\itemize{
  \item{\code{new}: Instantiates a new waypoint, arguements can be found at \code{\link{.init}}}	
  \item{\code{start}: Starts monitoring a waypoint.}	
  \item{\code{destroy}: Destroy a waypoint.}	
  \item{\code{enable}: Enable a waypoint.}	
  \item{\code{disable}: Disable awaypoint.}	
  \item{\code{get_direction}: Get the direction of the scroll.}	
  \item{\code{get_next}: Get the next waypoint.}	
  \item{\code{get_previous}: Get the previous waypoint.}	
}
}

\keyword{datasets}
