% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{.init}
\alias{.init}
\title{Initialise}
\usage{
.init(self, dom_id, offset = NULL, id = NULL)
}
\arguments{
\item{self}{A \code{waypoint} object.}

\item{dom_id}{Id of element to attach waypoint to.}

\item{offset}{Vertical offset, an integer (pixels) or a percentage passed as a string e.g. \code{"10"}.}

\item{id}{Id of waypoint.}
}
\description{
Initialise a waypoint
}
