% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{.init}
\alias{.init}
\title{Initialise}
\usage{
.init(self, dom_id, animate = FALSE, animation = "bounce",
  offset = NULL, horizontal = FALSE, id = NULL)
}
\arguments{
\item{self}{A \code{waypoint} object.}

\item{dom_id}{Id of element to attach waypoint to.}

\item{animate}{Whether to animate the anchor when the waypoint is triggered.}

\item{animation}{Animation to use, note that this can be overriden in the \code{animate} method.}

\item{offset}{Vertical offset, an integer (pixels) or a percentage passed as a string e.g. \code{"10"}.}

\item{horizontal}{Whether to apply waypoints of horizontal offset.}

\item{id}{Id of waypoint, see details.}
}
\description{
Initialise a waypoint
}
\details{
If you do not provide an \code{id} then you will have to rely on the \code{get_*} family of
of methods to get the callbacks. If you do, you can still rely on the latter but also may access 
the callbacks the traditional way with `input$idName_direction`, `input$idName_next`, and 
`input$idName_previous`.
}
\section{Animations}{

\itemize{
  \item{\code{bounce}} 
  \item{\code{flash}} 
  \item{\code{wobble}} 
  \item{\code{bounceInLeft}} 
  \item{\code{bounceOutDown}} 
  \item{\code{fadeIn}} 
  \item{\code{fadeInLeftBig}} 
  \item{\code{fadeInUpBig}} 
  \item{\code{flipOutX}} 
  \item{\code{fadeOutLeft}} 
  \item{\code{fadeOutUp}} 
  \item{\code{rotateIn}} 
  \item{\code{rotateInUpRight}} 
  \item{\code{rotateOutUpLeft}} 
  \item{\code{rollIn}}
  \item{\code{zoomInLeft}} 
  \item{\code{zoomOutDown}} 
  \item{\code{slideInDown}} 
  \item{\code{slideOutDown}} 
  \item{\code{heartBeat}} 
  \item{\code{flash}} 
  \item{\code{headShake}} 
  \item{\code{jello}} 
  \item{\code{bounceInRight}} 
  \item{\code{bounceOutLeft}}  
  \item{\code{fadeInDown}} 
  \item{\code{fadeInRight}} 
  \item{\code{fadeOut}} 
  \item{\code{fadeOutLeftBig}} 
  \item{\code{fadeOutUpBig}} 
  \item{\code{flipOutY}} 
  \item{\code{rotateInDownLeft}} 
  \item{\code{rotateOut}} 
  \item{\code{rotateOutUpRight}} 
  \item{\code{rollOut}} 
  \item{\code{zoomInRight}} 
  \item{\code{zoomOutLeft}} 
  \item{\code{slideInLeft}}  
  \item{\code{slideOutLeft}} 
  \item{\code{pulse}} 
  \item{\code{swing}} 
  \item{\code{bounceIn}} 
  \item{\code{bounceInUp}} 
  \item{\code{bounceOutRight}} 
  \item{\code{fadeInDownBig}} 
  \item{\code{fadeInRightBig}} 
  \item{\code{fadeOutDown}} 
  \item{\code{fadeOutRight}} 
  \item{\code{flipInX}} 
  \item{\code{lightSpeedIn}} 
  \item{\code{rotateInDownRight	}}  
  \item{\code{rotateOutDownLeft}} 
  \item{\code{hinge}} 
  \item{\code{zoomIn}} 
  \item{\code{zoomInUp}} 
  \item{\code{zoomOutRight}} 
  \item{\code{slideInRight	}} 
  \item{\code{slideOutRight}} 
  \item{\code{rubberBand}} 
  \item{\code{tada}} 
  \item{\code{bounceInDown}} 
  \item{\code{bounceOut}} 
  \item{\code{bounceOutUp}}  
  \item{\code{fadeInLeft}} 
  \item{\code{fadeInUp}} 
  \item{\code{fadeOutDownBig}} 
  \item{\code{fadeOutRightBig}} 
  \item{\code{flipInY}} 
  \item{\code{lightSpeedOut}} 
  \item{\code{rotateInUpLeft}} 
  \item{\code{rotateOutDownRight}} 
  \item{\code{jackInTheBox}} 
  \item{\code{zoomInDown}} 
  \item{\code{zoomOut}} 
  \item{\code{zoomOutUp}}  
  \item{\code{slideInUp}}  
  \item{\code{slideOutUp}}  
}
}

