% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waypoints.R
\name{waypoints}
\alias{waypoints}
\alias{Waypoint}
\alias{waypointer}
\alias{Waypointer}
\title{Waypoint}
\description{
A waypoint object to track.
}
\examples{
library(shiny)


ui <- fluidPage(
	use_waypointer(),
	div(
		h1("Scroll down"), 
		style = "min-height:90vh"
	),
	verbatimTextOutput("result"),
	plotOutput("plot"),
	div(style = "min-height:90vh")
)

server <- function(input, output, session) {

	w <- Waypoint$
		new("plot", offset = "20\%")$
		start()

	output$result <- renderPrint({
		w$get_direction()
	})

	output$plot <- renderPlot({

		req(w$get_direction())

		if(w$get_direction() == "down")
			hist(runif(100))
		else
			""
	})

}

if(interactive()) shinyApp(ui, server)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Waypoint$new()}}
\item \href{#method-start}{\code{Waypoint$start()}}
\item \href{#method-destroy}{\code{Waypoint$destroy()}}
\item \href{#method-enable}{\code{Waypoint$enable()}}
\item \href{#method-disable}{\code{Waypoint$disable()}}
\item \href{#method-animate}{\code{Waypoint$animate()}}
\item \href{#method-get_direction}{\code{Waypoint$get_direction()}}
\item \href{#method-going_up}{\code{Waypoint$going_up()}}
\item \href{#method-going_down}{\code{Waypoint$going_down()}}
\item \href{#method-get_triggered}{\code{Waypoint$get_triggered()}}
\item \href{#method-clone}{\code{Waypoint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waypoint$new(
  id,
  animate = FALSE,
  animation = "shake",
  offset = NULL,
  horizontal = FALSE,
  waypoint_id = NULL,
  start = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Id of element to use as waypoint.}

\item{\code{animate}}{Whether to animate element when the waypoint is triggered.}

\item{\code{animation}}{Animation to use if \code{animate} is set.}

\item{\code{offset}}{Offset relative to viewport to trigger the waypoint.}

\item{\code{horizontal}}{Set to \code{TRUE} if using horizontally.}

\item{\code{waypoint_id}}{Id of waypoint, useful to get the input value.}

\item{\code{start}}{Whether to automatically start watching the waypoint.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialise
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start"></a>}}
\subsection{Method \code{start()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waypoint$start()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Start watching the waypoint.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-destroy"></a>}}
\subsection{Method \code{destroy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waypoint$destroy()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Destroy the waypoint.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-enable"></a>}}
\subsection{Method \code{enable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waypoint$enable()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Enable the waypoint.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-disable"></a>}}
\subsection{Method \code{disable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waypoint$disable()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Disable the waypoint.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-animate"></a>}}
\subsection{Method \code{animate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waypoint$animate(animation = NULL)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Animate the waypoint.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_direction"></a>}}
\subsection{Method \code{get_direction()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waypoint$get_direction()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get direction in which user is scrolling past the waypoint
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-going_up"></a>}}
\subsection{Method \code{going_up()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waypoint$going_up()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Whether user is scrolling up past the waypoint.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-going_down"></a>}}
\subsection{Method \code{going_down()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waypoint$going_down()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Whether user is scrolling down past the waypoint.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_triggered"></a>}}
\subsection{Method \code{get_triggered()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waypoint$get_triggered()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Whether waypoint has been triggered.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waypoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
