% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waypoints.R
\docType{data}
\name{waypoints}
\alias{waypoints}
\alias{Waypoint}
\alias{waypointer}
\alias{Waypointer}
\title{Waypoint}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Waypoint
}
\description{
A waypoint object to track.
}
\section{Methods}{

\itemize{
  \item{\code{new}: Instantiates a new waypoint, arguements can be found at \code{\link{.init}}}	
  \item{\code{start}: Starts monitoring a waypoint.}	
  \item{\code{destroy}: Destroy a waypoint.}	
  \item{\code{enable}: Enable a waypoint (enabled by default).}	
  \item{\code{disable}: Disable a waypoint.}	
  \item{\code{animate}: Animate a waypoint, accepts, optionally, an animation.}	
  \item{\code{get_direction}: Returns the direction in which the user scrolls passed the waypoint.}	
  \item{\code{get_previous}: eturns `TRUE` if the waypoint has been triggered previously, and `FALSE` otherwise.}	
}
}

\examples{
library(shiny)


ui <- fluidPage(
	use_waypointer(),
	div(
		h1("Scroll down"), 
		style = "min-height:90vh"
	),
	verbatimTextOutput("result"),
	plotOutput("plot"),
	div(style = "min-height:90vh")
)

server <- function(input, output, session) {

	w <- Waypoint$
		new("plot", offset = "20\%")$
		start()

	output$result <- renderPrint({
		w$get_direction()
	})

	output$plot <- renderPlot({

		req(w$get_direction())

		if(w$get_direction() == "down")
			hist(runif(100))
		else
			""
	})

}

if(interactive()) shinyApp(ui, server)
}
\keyword{datasets}
